get_week_start to replace get_prior_monday
I ran into a problem with get_prior_monday, which is that it was counting
all Monday visits as belonging to the prior week (naturally). I only noticed
because I happened to run the report on a Monday - the output thought there
were two counts of appointments for the same week.

get_week_start is a mildly more sophisticated attempt. Let's see how it breaks!